[BACKPORT] Move CharacterComposer into //ui/base/ime
This patch makes CharacterComposer class reusable by other
platforms, by moving character_composer{.h,.cc} files and its
tests and auxiliary scripts out from ui/base/ime/chromeos
into ui/base/ime so that it can be reused by other platforms.

In particular, we intend to use it to properly handle dead keys
in ozone/wayland configuration. The ozone/wayland integration
will be done in a follow-up CL. In the future, ozone/x11 could
also benefit from this.

BUG=921947

Reviewed-on: https://chromium-review.googlesource.com/c/1446477
Commit-Queue: Nick Yamane <nickdiego@igalia.com>
Cr-Commit-Position: refs/heads/master@{#627991}
