+++ src/ui/base/ime/character_composer_sequences.txt
