+++ src/ui/base/ime/generate_character_composer_data.py
